Custom array field corrected, geoipdb functionality added.
Additionally, the index.html page now has aggregate statistics, though
some do not seem correct. A custom aggregate filter was added to get
the number of relays with a certain flag. A custom type cast was also
written for psycopg2, as psycopg2 throws away array indices when
converting SQL output to a python type. This makes using the default
type cast for 'BIGINT[]' impossible. This will be documented when a
proper location is decided on for the custom type cast.