Documented statusapp.models, lines <= 72char.
Also experimented with custom fields for the IP addresses: we're running
into a difficulty in which lte and gte are implemented
counterintuitively for IPAddressFields. Our solution now is to use just
a bit of raw SQL and string manipulation to access the information that
we need from these fields. While this implementation is efficient, it
is probably not good practice. So far, alternative implementations have
been slow, because custom and alternative fields used so far have not
taken advantage of the index on ipstart and ipend in the Tor Metrics
database.

Additionally, much of the documentation in statusapp.models is
guesswork, though after reading a good amount of Tor Metrics
documentation. It would be nice if someone more familiar with the
database could validate our documentation.