Wrote app logic in filters into views.
Not only does this make the application much easier to understand, but
it also highlights a fundamental problem with TorStatus as it exists
now: that data must be formatted properly before each use. TorStatus
needs persistent data storage for the data that is used so often. I
think that what is now "app logic" should be moved to the model level,
and our underlying database, whether it be Tor Metrics or a small
TorStatus cache, should store data in the same format that we will use
it in.